Check switchery element state before changing it
If an element is enabled and you call element.enable(),
the enable function checks the element status and if the action (enable)
and state (enabled) are the same it stops before doing anything or changing the element.
Otherwise enabling an already enabled element,
would break the functionality for this switchery element.
Same for disabling an element.